eza fix tab auto completion
If you have flag --icons but don't have the value it will tab complete
the value for you first which is not what you would expect when you tab
complete ls, l, la or lt, where you would expect file or directory
names. Solution is to always hard code value to it so it will never tab
complete for you.